+++ magicclass/_gui/_macro_utils.py
@@ -8,7 +8,6 @@ from magicgui.widgets.bases import ValueWidget
 from .utils import get_parameters
 from magicclass.utils import get_signature, thread_worker
 from magicclass.signature import MagicMethodSignature
-from magicclass.types import _StoredMeta
 
 if TYPE_CHECKING:
     from ._base import MagicTemplate
@@ -233,37 +232,6 @@ def _define_macro_recorder(sig: inspect.Signature, func: Callable):
                 ):
                     bgui.macro.pop()
                     bgui.macro._erase_last()
-
-            bgui.macro.append(expr)
-            bgui.macro._last_setval = None
-            return None
-
-    elif isinstance(sig.return_annotation, _StoredMeta):
-        if _auto_call and sig.return_annotation._maxsize == float("inf"):
-            raise ValueError(
-                f"Cannot use auto_call=True with a Stored type {sig.return_annotation}"
-                " with infinite size."
-            )
-
-        def _record_macro(bgui: MagicTemplate, out, *args, **kwargs):
-            bound = sig.bind(*args, **kwargs)
-            kwargs = dict(bound.arguments.items())
-            method_expr = Expr.parse_method(bgui, func, (), kwargs)
-            target = Symbol.asvar(out)
-            expr = Expr(Head.assign, [target, method_expr])
-
-            if _auto_call:
-                last_expr = bgui.macro[-1]
-                if last_expr.head is Head.assign and last_expr[0] == target:
-                    last_method_expr = last_expr.args[1]
-                    if (
-                        last_method_expr.head is Head.call
-                        and last_method_expr.args[0].head is Head.getattr
-                        and last_method_expr.at(0, 1) == expr.at(0, 1)
-                        and len(bgui.macro) > 0
-                    ):
-                        bgui.macro.pop()
-                        bgui.macro._erase_last()
 
             bgui.macro.append(expr)
             bgui.macro._last_setval = None
