+++ magicclass/_register_types.py
@@ -4,7 +4,7 @@ from macrokit import Expr, Head, symbol
 from enum import Enum
 import pathlib
 import datetime
-from magicclass.types import Color, Stored, Path, ExprStr
+from magicclass.types import Color, Path, ExprStr
 from magicclass.widgets import ColorEdit, EvalLineEdit
 from magicclass._gui._base import MagicTemplate
 
@@ -57,12 +57,6 @@ def find_myname(gui: MagicTemplate):
 import magicgui as mgui
 
 mgui.register_type(Color, widget_type=ColorEdit)
-mgui.register_type(
-    Stored,
-    choices=Stored._get_choice,
-    return_callback=Stored._store_value,
-)
-
 mgui.register_type(Path.Save, widget_type="FileEdit", mode="w")
 mgui.register_type(Path.Dir, widget_type="FileEdit", mode="d")
 mgui.register_type(Path.Multiple, widget_type="FileEdit", mode="rm")
