+++ tests/test_types.py
@@ -2,7 +2,7 @@ from typing_extensions import Annotated
 
 import pytest
 from magicclass import magicclass, set_options, get_function_gui
-from magicclass.types import Optional, Union, Stored, Path
+from magicclass.types import Optional, Union, Path
 from magicclass import widgets
 
 def test_basics():
@@ -90,73 +90,6 @@ def test_union_with_set_option():
     assert wdt[1].widget_type == "FloatSpinBox"
     assert wdt[0].max == 4
     assert wdt[1].max == 5.0
-
-def test_stored_type():
-    class X:
-        def __init__(self, val):
-            self.val = val
-
-        def __eq__(self, o: object) -> bool:
-            if type(o) is not X:
-                return False
-            return self.val == o.val
-
-        def __repr__(self):
-            return f"X({self.val!r})"
-
-    @magicclass
-    class A:
-        def provide(self, x: int) -> Stored[X]:
-            return X(x)
-
-        def receive(self, s: Stored[X]):
-            pass
-
-    ui = A()
-    provide = get_function_gui(ui.provide)
-    receive = get_function_gui(ui.receive)
-    provide(2)
-    assert receive.s.choices == (X(2),)
-    provide(2)
-    assert receive.s.choices == (X(2), X(2))
-    provide(5)
-    # NOTE: ComboBox sometimes does not keep the order...
-    # assert receive.s.choices == (X(5), X(2), X(2))
-    receive()
-    lefts = [str(ui.macro[i]).split("=")[0].strip() for i in (1, 2, 3)]
-    assert lefts[0] != lefts[1] and lefts[1] != lefts[2]
-    assert str(ui.macro[4]) == f"ui.receive(s={lefts[-1]})"
-    assert X(3) == ui.provide(3), "provide did not work programmatically"
-
-def test_stored_last_type():
-    class X:
-        def __init__(self, val):
-            self.val = val
-
-        def __eq__(self, o: object) -> bool:
-            if type(o) is not X:
-                return False
-            return self.val == o.val
-
-        def __repr__(self):
-            return f"X({self.val!r})"
-
-    @magicclass
-    class A:
-        def provide(self, x: int) -> Stored[X]:
-            return X(x)
-
-        def receive(self, s: Stored.Last[X]):
-            pass
-
-
-    ui = A()
-    provide = get_function_gui(ui.provide)
-    receive = get_function_gui(ui.receive)
-    provide(2)
-    assert receive.s.value == X(2)
-    provide(3)
-    assert receive.s.value == X(3)
 
 @pytest.mark.parametrize(
     "typ, mode",
